pyproject.toml
[tool.poetry]
name = "building-blocks"
version = "0.1.0"
description = "A clean architecture Python library following hexagonal principles"
authors = ["Glauber Brennon <glauberbrennon@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "building_blocks", from = "src"}]
keywords = [
    "clean architecture",
    "hexagonal architecture",
    "python library",
    "framework agnostic",
    "domain-driven design",
    "ddd",
    "software architecture",
    "software design",
    "software engineering",
    "framework-agnostic",
]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
mypy = "1.16.1"
bandit = "*"
pre-commit = "4.2.0"
pytest = "*"
pytest-cov = "*"
pytest-asyncio = "*"
ruff = "0.12.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
fastapi = {extras = ["standard"], version = "^0.115.13"}
httpx = "^0.28.1"
pydantic-settings = "^2.10.0"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.16.3"
black = "^25.1.0"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests", "examples", "scripts"]
fix = true
exclude = [
    "**/migrations/**",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
]
ignore = [
    "D203",
    "D213",
    "B008"
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
ignore_missing_imports = true
disallow_untyped_defs = false

[tool.bandit]
exclude_dirs = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=building_blocks",
    "--cov-report=term-missing",
]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "src/building_blocks/application/ports/**/*.py",
    "src/building_blocks/domain/ports/**/*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError"
]
